Refactor project structure into modules with debug logging
