+++ bot/constants.py
@@ -1,0 +1,20 @@
+# Conversation states
+(
+    COMMENT,
+    ADD_TASK_TITLE,
+    ADD_TASK_CATEGORY_CHOOSE,
+    ADD_TASK_CATEGORY_INPUT,
+    ADD_TASK_PRIORITY,
+    ADD_TASK_TAGS,
+    EDIT_TASK_TITLE,
+    EDIT_TASK_CATEGORY_CHOOSE,
+    EDIT_TASK_CATEGORY_INPUT,
+    EDIT_TASK_PRIORITY,
+    EDIT_TASK_TAGS,
+    CATEGORY_MENU,
+    CATEGORY_EDIT,
+    CATEGORY_ADD,
+    SETTINGS_MENU,
+    SETTINGS_TIME,
+    FILTER_MENU,
+) = range(17)
